+++ src/lib.rs
@@ -107,7 +107,7 @@
 //!
 //!     let group = client
 //!         // Initialise up to 16 slave devices
-//!         .init::<MAX_SLAVES, _>(group, |groups, slave| groups.push(slave))
+//!         .init::<16, _>(group, |groups, _slave| Ok(groups.as_mut()))
 //!         .await
 //!         .expect("Init");
 //!